+++ NEWS.md
@@ -1,5 +1,6 @@
 # hlatools 0.0.6.9000
 
+* [feature] `db_version()` returns a `numeric_version` object instead of a `character` string (#3).
 * [feature] new `hlatools_version()` method for HLAGene objects (#2).
 * [bug] fix `locusname()` method for HLAGene objects (#1).
 * [feature] add `exon()`, `intron()`, and `utr()` to extract feature sequences